# src/generated/resources/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Confluent ide-sidecar API",
    "description" : "API for the Confluent ide-sidecar, part of Confluent for VS Code",
    "termsOfService" : "Your terms here",
    "contact" : {
      "name" : "Confluent for VS Code Support",
      "url" : "https://confluent.io/contact",
      "email" : "vscode@confluent.io"
    },
    "version" : "0.172.0"
  },
  "servers" : [ {
    "url" : "http://127.0.0.1:26636",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/gateway/v1/callback-vscode-docs" : {
      "get" : {
        "tags" : [ "O Auth Callback Resource" ],
        "parameters" : [ {
          "name" : "code",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "state",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}